Refactored the networking code out of the AbstractVC
